updated scripts with new validated pipeline
