+++ ai_genomics/getters/patents.py
@@ -10,7 +10,6 @@ def get_ai_genomics_patents() -> pd.DataFrame:
         - application_number
         - publication_number
         - full list of cpc codes
-        - full list of ipc codes
         - abstract_text
         - publication_date
         - inventor
@@ -18,7 +17,7 @@ def get_ai_genomics_patents() -> pd.DataFrame:
     """
     return load_s3_data(
         bucket_name,
-        "inputs/patent_data/processed_patent_data/ai_genomics_patents_cpc_ipc_codes.csv",
+        "inputs/patent_data/processed_patent_data/ai_genomics_patents_cpc_codes.csv",
     )
 
 
@@ -28,7 +27,6 @@ def get_ai_sample_patents() -> pd.DataFrame:
         - application_number
         - publication_number
         - full list of cpc codes
-        - full list of ipc codes
         - abstract_text
         - publication_date
         - inventor
@@ -36,7 +34,7 @@ def get_ai_sample_patents() -> pd.DataFrame:
     """
     return load_s3_data(
         bucket_name,
-        "inputs/patent_data/processed_patent_data/ai_sample_patents_cpc_ipc_codes.csv",
+        "inputs/patent_data/processed_patent_data/ai_sample_patents_cpc_codes.csv",
     )
 
 
@@ -46,7 +44,6 @@ def get_genomics_sample_patents() -> pd.DataFrame:
         - application_number
         - publication_number
         - full list of cpc codes
-        - full list of ipc codes
         - abstract_text
         - publication_date
         - inventor
@@ -54,7 +51,7 @@ def get_genomics_sample_patents() -> pd.DataFrame:
     """
     return load_s3_data(
         bucket_name,
-        "inputs/patent_data/processed_patent_data/genomics_sample_patents_cpc_ipc_codes.csv",
+        "inputs/patent_data/processed_patent_data/genomics_sample_patents_cpc_codes.csv",
     )
 
 
@@ -63,24 +60,6 @@ def get_ai_genomics_cpc_codes() -> Mapping[str, Mapping[str, Union[str, str]]]:
     return load_s3_data(
         bucket_name, "outputs/patent_data/class_codes/cpc_with_definitions.json"
     )
-
-
-def get_ai_genomics_ipc_codes() -> Mapping[str, Mapping[str, Union[str, str]]]:
-    """From S3 loads AI in genomics ipc codes WITHOUT Google BigQuery ipc code formatting"""
-    return load_s3_data(
-        bucket_name, "outputs/patent_data/class_codes/ipc_with_definitions.json"
-    )
-
-
-def get_ai_genomics_ipc_codes_formatted() -> Mapping[
-    str, Mapping[str, Union[str, str]]
-]:
-    """From S3 loads AI in genomics ipc codes WITH Google BigQuery ipc code formatting"""
-    return load_s3_data(
-        bucket_name,
-        "outputs/patent_data/class_codes/ipc_formatted_with_definitions.json",
-    )
-
 
 def get_ai_genomics_patents_entities() -> Mapping[str, Mapping[str, Union[str, str]]]:
     """From S3 loads AI in genomics patents entities"""
