+++ ai_genomics/pipeline/patent_data/get_ai_genomics_patents.py
@@ -1,8 +1,7 @@
-"""Script to query BigQuery based on genomics related and AI related cpc/ipc codes."""
+"""Script to query BigQuery based on genomics related and AI related cpc codes."""
 from ai_genomics import bucket_name, logger
 from ai_genomics.getters.patents import (
     get_ai_genomics_cpc_codes,
-    get_ai_genomics_ipc_codes_formatted,
 )
 from ai_genomics.utils.patents import (
     est_conn,
@@ -15,7 +14,7 @@ from typing import Dict
 import pandas as pd
 
 GENOMICS_AI_FIELDS = (
-    "publication_number, application_number, cpc.code as cpc_code, ipc.code as ipc_code, "
+    "publication_number, application_number, cpc.code as cpc_code, "
     "title_localized.text as title_text, title_localized.language as title_language, "
     "abstract_localized.text as abstract_text, abstract_localized.language as abstract_language, "
     "publication_date, filing_date, grant_date, priority_date, inventor_harmonized.name "
@@ -25,83 +24,64 @@ GENOMICS_AI_FIELDS = (
 )
 
 CPC_CODES = get_ai_genomics_cpc_codes()
-IPC_CODES = get_ai_genomics_ipc_codes_formatted()
-
 
 def code_query(
     cpc_codes: Dict[str, list] = CPC_CODES,
-    ipc_codes: Dict[str, list] = IPC_CODES,
     topic: str = "ai",
     sample: int = 10,
 ) -> str:
-    """Generates query to get sample of patents based on cpc/ipc codes.
+    """Generates query to get sample of patents based on cpc codes.
 
     Args:
         cpc_codes: dictionary of cpc codes.
-        ipc_codes: dictionary of ipc codes.
 
     Returns:
         BigQuery query to select related patents.
     """
-    cpc_ids, ipc_ids = (
-        convert_list_of_codes_to_string(list(cpc_codes[topic].keys())),
-        convert_list_of_codes_to_string(list(ipc_codes[topic].keys())),
-    )
+    cpc_ids = convert_list_of_codes_to_string(list(cpc_codes[topic].keys())),
 
     topic_q = (
         f"SELECT {GENOMICS_AI_FIELDS} "
         f"FROM `patents-public-data.patents.publications` TABLESAMPLE SYSTEM ({sample} PERCENT), "
-        " UNNEST(cpc) AS cpc, UNNEST(ipc) AS ipc, "
+        " UNNEST(cpc) AS cpc, "
         "UNNEST(title_localized) AS title_localized, UNNEST(abstract_localized) AS abstract_localized, "
         "UNNEST(inventor_harmonized) inventor_harmonized, UNNEST(assignee_harmonized) assignee_harmonized "
-        f"WHERE cpc.code IN ({cpc_ids}) OR "
-        f"ipc.code IN ({ipc_ids})"
-    )
+        f"WHERE cpc.code IN ({cpc_ids})")
 
     return topic_q
 
 
 def genomics_ai_query(
     cpc_codes: Dict[str, list] = CPC_CODES,
-    ipc_codes: Dict[str, list] = IPC_CODES,
 ) -> str:
     """Generates query to identify genomics ai patents
-            based on cpc and ipc codes.
+            based on cpc codes.
 
     Args:
         cpc_codes: dictionary of genomics and ai related cpc codes.
-        ipc_codes: dictionary of genomics and ai related ipc codes.
 
     Returns:
         BigQuery query to select genomics and ai related patents.
     """
 
-    cpc_ai_ids, ipc_ai_ids = (
-        convert_list_of_codes_to_string(list(cpc_codes["ai"].keys())),
-        convert_list_of_codes_to_string(list(ipc_codes["ai"].keys())),
-    )
-    cpc_genomics_ids, ipc_genomics_ids = (
-        convert_list_of_codes_to_string(list(cpc_codes["genomics"].keys())),
-        convert_list_of_codes_to_string(list(ipc_codes["genomics"].keys())),
-    )
+    cpc_ai_ids = convert_list_of_codes_to_string(list(cpc_codes["ai"].keys()))
+    cpc_genomics_ids = convert_list_of_codes_to_string(list(cpc_codes["genomics"].keys())),
 
     genomics_q = (
         f"SELECT DISTINCT publication_number "
-        "FROM `patents-public-data.patents.publications`, UNNEST(cpc) AS cpc__u, UNNEST(ipc) AS ipc__u "
-        f"WHERE cpc__u.code IN ({cpc_genomics_ids}) OR "
-        f"ipc__u.code IN ({ipc_genomics_ids})"
-    )
+        "FROM `patents-public-data.patents.publications`, UNNEST(cpc) AS cpc__u, "
+        f"WHERE cpc__u.code IN ({cpc_genomics_ids})")
 
     genomics_ai_q = (
         f"WITH "
         f"genomics_ids AS ({genomics_q}) "
         "SELECT "
         f"{GENOMICS_AI_FIELDS} "
-        "FROM `patents-public-data.patents.publications`, UNNEST(cpc) AS cpc, UNNEST(ipc) AS ipc, "
+        "FROM `patents-public-data.patents.publications`, UNNEST(cpc) AS cpc, "
         "UNNEST(title_localized) AS title_localized, UNNEST(abstract_localized) AS abstract_localized, "
         "UNNEST(inventor_harmonized) inventor_harmonized, UNNEST(assignee_harmonized) assignee_harmonized "
         "INNER JOIN genomics_ids USING(publication_number) "
-        f"WHERE cpc.code in ({cpc_ai_ids}) OR ipc.code in ({ipc_ai_ids});"
+        f"WHERE cpc.code in ({cpc_ai_ids})"
     )
 
     return genomics_ai_q
@@ -157,5 +137,5 @@ if __name__ == "__main__":
         ("ai_genomics", "ai_sample", "genomics_sample"),
         (ai_genomics_patents, ai_patents_sample, genomics_patents_sample),
     ):
-        full_table_name = f"inputs/patent_data/processed_patent_data/{table_name}_patents_cpc_ipc_codes.csv"
+        full_table_name = f"inputs/patent_data/processed_patent_data/{table_name}_patents_cpc_codes.csv"
         save_to_s3(bucket_name, table, full_table_name)
